+++ CHANGES.rst
@@ -4,15 +4,15 @@ Products.MeetingCharleroi Changelog
 4.1.9 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Fixed `test_ValidateCategoryIfCollegeItemToSendToCouncil` due to change in
+  `MeetingItem.validate_category`, the given category must exists.
+  [gbastien]
 
 4.1.8 (2021-07-26)
 ------------------
 
 - Added a default classifier when an item is sent from one MeetingConfig to another.
   [aduchene]
-
 
 4.1.7 (2020-10-30)
 ------------------